Fix T73080: Remove support for scene radius in Weight/Vertex paint
The 2D paint cursor in the 3D view does not support scene radius in
weight and vertex paint mode. This was also the case in 2.80 with the
old cursor. The option to change the units from view to scene is not
available in the UI in those modes, so I think it makes sense to remove
the support from the code.

The ideal solution could be supporting the new cursor in all paint
modes, but that is going to take more time and it is going to be
problematic in texture paint. After doing this, we can bring this code
back and add the option to change the units to the UI.

Reviewed By: jbakker

Maniphest Tasks: T73080

Differential Revision: https://developer.blender.org/D6586